Renamed eigenvalue/k_effective to eigenvalue/k_eigenvalue
