+++ src/eigenvalue/k_eigenvalue/k_effective_updater_i.hpp
@@ -1,5 +1,5 @@
-#ifndef BART_SRC_EIGENVALUE_K_EFFECTIVE_K_EFFECTIVE_UPDATER_I_HPP_
-#define BART_SRC_EIGENVALUE_K_EFFECTIVE_K_EFFECTIVE_UPDATER_I_HPP_
+#ifndef BART_SRC_EIGENVALUE_K_EIGENVALUE_K_EFFECTIVE_UPDATER_I_HPP_
+#define BART_SRC_EIGENVALUE_K_EIGENVALUE_K_EFFECTIVE_UPDATER_I_HPP_
 
 #include <optional>
 
@@ -18,4 +18,4 @@ class K_EffectiveUpdaterI : public utility::HasDescription {
 
 } // namespace bart::eigenvalue::k_effective
 
-#endif // BART_SRC_EIGENVALUE_K_EFFECTIVE_K_EFFECTIVE_UPDATER_I_HPP_
+#endif // BART_SRC_EIGENVALUE_K_EIGENVALUE_K_EFFECTIVE_UPDATER_I_HPP_
