+++ src/framework/builder/tests/framework_builder_integration_tests.cpp
@@ -13,8 +13,8 @@
 #include "data/cross_sections/material_cross_sections.hpp"
 #include "domain/finite_element/finite_element_gaussian.hpp"
 #include "domain/domain.hpp"
-#include "eigenvalue/k_effective/updater_via_fission_source.h"
-#include "eigenvalue/k_effective/updater_via_rayleigh_quotient.hpp"
+#include "eigenvalue/k_eigenvalue/updater_via_fission_source.h"
+#include "eigenvalue/k_eigenvalue/updater_via_rayleigh_quotient.hpp"
 #include "formulation/scalar/diffusion.h"
 #include "formulation/scalar/drift_diffusion.hpp"
 #include "formulation/angular/self_adjoint_angular_flux.h"
@@ -46,7 +46,7 @@
 #include "convergence/tests/iteration_completion_checker_mock.hpp"
 #include "domain/tests/domain_mock.hpp"
 #include "domain/finite_element/tests/finite_element_mock.hpp"
-#include "eigenvalue/k_effective/tests/k_effective_updater_mock.h"
+#include "eigenvalue/k_eigenvalue/tests/k_effective_updater_mock.h"
 #include "formulation/angular/tests/self_adjoint_angular_flux_mock.h"
 #include "formulation/scalar/tests/diffusion_mock.h"
 #include "formulation/scalar/tests/drift_diffusion_mock.hpp"
